Add some switches so that only parts of the configuration can be applied.

+++ scripts/vmware/Set-VMHostSettingsForJMU.ps1
@@ -11,7 +11,19 @@ param (
         $VSphereCredential,
 
         [Parameter(Mandatory=$true)]
-        $HostCredential
+        $HostCredential,
+
+        [switch]
+        $DellOMCheck,
+
+        [switch]
+        $ConfigureNTP=$true,
+
+        [switch]
+        $ConfigureSyslog=$true,
+
+        [switch]
+        $ConfigureSNMP=$true
       )
 
 function WriteWarning($message) {
@@ -45,41 +57,42 @@ if ($vSphereServer -eq $null) {
     return
 }
 
-$server = Get-VMHost $VMHost
+$server = @(Get-VMHost $VMHost)[0]
 if ($server -eq $null) {
     Write-Error "Could not find host $VMHost"
     return
 }
 
 # Check for OpenManage
-$patches = Get-VMHostPatch -Server $vSphereServer -VMHost $Server
-$foundOM = $false
-$foundDellImage = $false
+if ($DellOMCheck) {
+    $patches = Get-VMHostPatch -Server $vSphereServer -VMHost $Server
+    $foundOM = $false
+    $foundDellImage = $false
 
-WriteHost "Updating esxupdate.conf"
-$svc = Get-VMHostService $server | ? { $_.Key -eq 'TSM-SSH' }
-if ($svc.Running -eq $false) {
-    $svc | Start-VMHostService
-    Start-Sleep 5
-}
-& 'C:\Program Files (x86)\PuTTY\plink.exe' -l $HostCredential.GetNetworkCredential().UserName -pw $HostCredential.GetNetworkCredential().Password $server "sed -re 's|file = ?$|file = /var/tmp/esxupdate.debug|' -i /etc/vmware/esxupdate/esxupdate.conf"
-
-foreach ($patch in $patches) {
-    if ($patch.Id -match "OpenManage") {
-        WriteHost "Found installed patch $($patch.Id) ($($patch.Description))"
-        $foundOM = $true
-    } elseif ($patch.Id -eq 'Dell') {
-        WriteHost "Dell Customized ESXi image found"
-        $foundDellImage = $true
+    WriteHost "Updating esxupdate.conf"
+    $svc = Get-VMHostService $server | ? { $_.Key -eq 'TSM-SSH' }
+    if ($svc.Running -eq $false) {
+        $svc | Start-VMHostService
+        Start-Sleep 5
     }
-}
+    & 'C:\Program Files (x86)\PuTTY\plink.exe' -l $HostCredential.GetNetworkCredential().UserName -pw $HostCredential.GetNetworkCredential().Password $server "sed -re 's|file = ?$|file = /var/tmp/esxupdate.debug|' -i /etc/vmware/esxupdate/esxupdate.conf"
 
-if ($foundDellImage -eq $false) {
-    WriteWarning "host was not built using the Dell-customized ESXi image"
-}
+    foreach ($patch in $patches) {
+        if ($patch.Id -match "OpenManage") {
+            WriteHost "Found installed patch $($patch.Id) ($($patch.Description))"
+            $foundOM = $true
+        } elseif ($patch.Id -eq 'Dell') {
+            WriteHost "Dell Customized ESXi image found"
+            $foundDellImage = $true
+        }
+    }
 
-if ($foundOM -eq $false) {
-    WriteWarning "Dell OpenManage not installed.  Please use VMware Update Manager to install."
+    if ($foundDellImage -eq $false) {
+        WriteWarning "host was not built using the Dell-customized ESXi image"
+    }
+
+    if ($foundOM -eq $false) {
+        WriteWarning "Dell OpenManage not installed.  Please use VMware Update Manager to install."
 #    $ds = Get-Datastore -Server $vSphereServer | ? { $_.Name -eq 'NAS:virtcluster-Mass-utility01' }
 #    if ($ds -eq $null) {
 #        WriteError "Could not find NAS:virtcluster-Mass-utility01 iSCSI LUN!"
@@ -91,31 +104,39 @@ if ($foundOM -eq $false) {
 #        WriteHost "Enabling CIM OEM Providers"
 #        Set-VMHostAdvancedConfiguration -Server $vSphereServer -VMHost $server -Name UserVars.CIMoemProviderEnabled -Value 1
 #    }
-} #else {
-    WriteHost "Enabling CIM OEM Providers"
-    Set-VMHostAdvancedConfiguration -Server $vSphereServer -VMHost $server -Name UserVars.CIMoemProviderEnabled -Value 1
-#}
-
-$ntpServers = @(Get-VMHostNtpServer -Server $vSphereServer -VMHost $server)
-foreach ($ntpServer in $ntpServers) {
-    WriteHost "Removing existing NTP Server $ntpServer"
-    Remove-VmHostNtpServer -Server $vSphereServer -VMHost $server -NtpServer $ntpServer -Confirm:$false
+    } #else {
+        WriteHost "Enabling CIM OEM Providers"
+        $null = Set-VMHostAdvancedConfiguration -Server $vSphereServer -VMHost $server -Name UserVars.CIMoemProviderEnabled -Value 1
+#   }
 }
-$gateway = (Get-VMHostNetwork -Server $vSphereServer -VMHost $server).VMKernelGateway
-WriteHost "Adding NTP Server $gateway"
-$null = Add-VMHostNtpServer -Server $vSphereServer -VMHost $server -NtpServer $gateway
 
-WriteHost "Enabling syslog to it-lms"
-$null = Set-VMHostSysLogServer -Server $vSphereServer -VMHost $server -SysLogServer it-lms.jmu.edu -SysLogServerPort 514
-
-WriteHost "Enabling SNMP"
-$null = Connect-VIServer -Server $server -Credential $HostCredential
-$hostSnmp = Get-VMHostSnmp
-foreach ($trapTarget in $hostSnmp.TrapTargets) {
-    $hostSnmp = Set-VMHostSnmp -HostSnmp $hostSnmp -TrapTargetToRemove $trapTarget
+if ($ConfigureNTP) {
+    $ntpServers = @(Get-VMHostNtpServer -Server $vSphereServer -VMHost $server)
+    foreach ($ntpServer in $ntpServers) {
+        WriteHost "Removing existing NTP Server $ntpServer"
+        Remove-VmHostNtpServer -Server $vSphereServer -VMHost $server -NtpServer $ntpServer -Confirm:$false
+    }
+    $gateway = (Get-VMHostNetwork -Server $vSphereServer -VMHost $server).VMKernelGateway
+    WriteHost "Adding NTP Server $gateway"
+    $null = Add-VMHostNtpServer -Server $vSphereServer -VMHost $server -NtpServer $gateway
 }
-$hostSnmp = Set-VMHostSnmp -HostSnmp $hostSnmp -ReadOnlyCommunity @("zenosscmnty", "hostwatch") 
-$hostSnmp = Set-VMHostSnmp -HostSnmp $hostSnmp -AddTarget -TargetCommunity "zenosscmnty" -TargetHost "it-zenoss1.jmu.edu" -TargetPort 161
-$hostSnmp = Set-VMHostSnmp -HostSnmp $hostSnmp -AddTarget -TargetCommunity "hostwatch" -TargetHost "ita.ad.jmu.edu" -TargetPort 161
-$hostSnmp = Set-VMHostSnmp -HostSnmp $hostSnmp -Enabled $true
+
+if ($ConfigureSyslog) {
+    WriteHost "Enabling syslog to it-lms"
+    $null = Set-VMHostSysLogServer -Server $vSphereServer -VMHost $server -SysLogServer it-lms.jmu.edu -SysLogServerPort 514
+}
+
+if ($ConfigureSNMP) {
+    WriteHost "Enabling SNMP"
+    $null = Connect-VIServer -Server $server -Credential $HostCredential
+    $hostSnmp = Get-VMHostSnmp
+    foreach ($trapTarget in $hostSnmp.TrapTargets) {
+        $hostSnmp = Set-VMHostSnmp -HostSnmp $hostSnmp -TrapTargetToRemove $trapTarget
+    }
+    $hostSnmp = Set-VMHostSnmp -HostSnmp $hostSnmp -ReadOnlyCommunity @("zenosscmnty", "hostwatch") 
+    $hostSnmp = Set-VMHostSnmp -HostSnmp $hostSnmp -AddTarget -TargetCommunity "zenosscmnty" -TargetHost "it-zenoss1.jmu.edu" -TargetPort 161
+    $hostSnmp = Set-VMHostSnmp -HostSnmp $hostSnmp -AddTarget -TargetCommunity "hostwatch" -TargetHost "ita.ad.jmu.edu" -TargetPort 161
+    $hostSnmp = Set-VMHostSnmp -HostSnmp $hostSnmp -AddTarget -TargetCommunity "hostwatch" -TargetHost "it-dome.ad.jmu.edu" -TargetPort 161
+    $hostSnmp = Set-VMHostSnmp -HostSnmp $hostSnmp -Enabled $true
+}
